requirements.txt
absl-py
addict
alembic
anyio
appdirs
argon2-cffi
async-generator
attrs
autopep8
Babel
backcall
black
bleach
cachetools
calmsize
certifi
cffi
chardet
click
cliff
cmaes
cmd2
colorama
colorlog
commonmark
cycler
decorator
defusedxml
dill
efficientnet-pytorch
entrypoints
gitdb
GitPython
google-auth
google-auth-oauthlib
graphviz
grpcio
idna
importlib-metadata
jedi
Jinja2
joblib
json5
jsonschema
jstyleson
jupyter
jupyter-client
jupyter-console
jupyter-core
jupyter-server
jupyterlab
jupyterlab-pygments
jupyterlab-server
jupyterlab-widgets
kiwisolver
line-profiler
Mako
Markdown
MarkupSafe
matplotlib
mistune
multiprocess
munch
mypy-extensions
mysqlclient
natsort
nbclassic
nbclient
nbconvert
nbformat
nest-asyncio
networkx
ninja
nncf
notebook
numpy
oauthlib
omegaconf
openvino
optuna
packaging
pandas
pandocfilters
parso
pathos
pathspec
pbr
pexpect
pickleshare
Pillow
pox
ppft
pretrainedmodels
prettytable
prometheus-client
prompt-toolkit
protobuf
ptyprocess
pyasn1
pyasn1-modules
pycodestyle
pycparser
pydot
Pygments
pylibjpeg-libjpeg
pyparsing
pyperclip
pyrsistent
python-dateutil
python-editor
pytorch-memlab
pytz
PyYAML
pyzmq
qtconsole
QtPy
regex
requests
requests-oauthlib
rich
rsa
scikit-learn
scipy
segmentation-models-pytorch
Send2Trash
six
smmap
sniffio
SQLAlchemy
stevedore
tensorboard
tensorboard-plugin-wit
terminado
testpath
texttable
threadpoolctl
timm
toml
tomli
torch==1.8.1
torch-tb-profiler
torchaudio
torchinfo
torchvision
tornado
tqdm
traitlets
typed-ast
typing-extensions
urllib3
wcwidth
webencodings
Werkzeug
widgetsnbextension
zipp
SimpleITK
